design java Datasource
