+++ DataSource/appsettings.json
@@ -1,5 +1,3 @@
 {
-    "Langues":"",
-    "ServiceName": "",
     "PackageName": ""
 }
